Add LabeledExamples class
